// src/app/shared/models/common_model.ts
export interface GridFilter {
  ColumnName: string;
  SortColumnName: string;
  DisplayText: string;
  CurrencyName: string;
  Value: any;
  Condition: string;
  Type: string; // string|date|datetime|int|number|decimal|bool
  Is_Visible: boolean;
  Width: number;
  TextAlign: string;
  Is_Sum: boolean;
  Is_Price: boolean;
  Is_Sort: boolean;
  Is_EditLink: boolean;
  EditType: string;
  Actions: Array<Action_Type>;
  Badges: Array<Badge_Type>;
  Is_TDClass: boolean;
  Is_Decimal: boolean;
}
export class Action_Type {
  class: string;
  text: string;
  font: string;
  type: string;
  tooltip?: string;
}

export class Badge_Type {
  text: string;
  condition: string;
}